+++ tests/unit/client/test_experiment_client.py
@@ -31,6 +31,15 @@ def test_properties(project_client):
     assert experiment.project == project
 
 
+def test_get_identifiers(project_client):
+    project = project_client
+    experiment = project.log_experiment()
+    project_name, experiment_id = experiment._get_parent_identifiers()
+
+    assert project_name == project.name
+    assert experiment_id == experiment.id
+
+
 def test_log_metric(project_client):
     project = project_client
     experiment = project.log_experiment(name="exp1")
